requirements.txt
djangorestframework
djangorestframework-simplejwt
coverage
Pillow
django-cors-headers